+++ crates/objc2/src/ffi/method.rs
@@ -3,16 +3,9 @@ use std::os::raw::c_char;
 use std::os::raw::c_uint;
 
 use crate::{
-    ffi::{objc_selector, OpaqueData},
-    runtime::Imp,
+    ffi::objc_selector,
+    runtime::{Imp, Method},
 };
-
-/// A type that represents a method in a class definition.
-#[repr(C)]
-pub struct objc_method {
-    _priv: [u8; 0],
-    _p: OpaqueData,
-}
 
 /// Describes an Objective-C method.
 #[repr(C)]
@@ -27,31 +20,31 @@ pub struct objc_method_description {
 extern_c! {
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The return value is deallocated with [`free`][crate::ffi::free].
-    pub fn method_copyArgumentType(method: *const objc_method, index: c_uint) -> *mut c_char;
+    pub fn method_copyArgumentType(method: *const Method, index: c_uint) -> *mut c_char;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The return value is deallocated with [`free`][crate::ffi::free].
-    pub fn method_copyReturnType(method: *const objc_method) -> *mut c_char;
+    pub fn method_copyReturnType(method: *const Method) -> *mut c_char;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_exchangeImplementations(method1: *mut objc_method, method2: *mut objc_method);
+    pub fn method_exchangeImplementations(method1: *mut Method, method2: *mut Method);
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     pub fn method_getArgumentType(
-        method: *const objc_method,
+        method: *const Method,
         index: c_uint,
         dst: *mut c_char,
         dst_len: usize,
     );
     #[cfg(any(doc, target_vendor = "apple"))]
-    pub fn method_getDescription(m: *const objc_method) -> *const objc_method_description;
+    pub fn method_getDescription(m: *const Method) -> *const objc_method_description;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_getImplementation(method: *const objc_method) -> Option<Imp>;
+    pub fn method_getImplementation(method: *const Method) -> Option<Imp>;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_getName(method: *const objc_method) -> *const objc_selector;
+    pub fn method_getName(method: *const Method) -> *const objc_selector;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_getNumberOfArguments(method: *const objc_method) -> c_uint;
+    pub fn method_getNumberOfArguments(method: *const Method) -> c_uint;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_getReturnType(method: *const objc_method, dst: *mut c_char, dst_len: usize);
+    pub fn method_getReturnType(method: *const Method, dst: *mut c_char, dst_len: usize);
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_getTypeEncoding(method: *const objc_method) -> *const c_char;
+    pub fn method_getTypeEncoding(method: *const Method) -> *const c_char;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn method_setImplementation(method: *const objc_method, imp: Imp) -> Option<Imp>;
+    pub fn method_setImplementation(method: *const Method, imp: Imp) -> Option<Imp>;
 }