+++ crates/objc2/src/ffi/mod.rs
@@ -282,6 +282,9 @@ pub type objc_object = crate::runtime::AnyObject;
 #[deprecated = "merged with `runtime::Imp`, and made non-null"]
 pub type IMP = Option<crate::runtime::Imp>;
 
+#[deprecated = "merged with `runtime::Imp`"]
+pub type objc_method = crate::runtime::Method;
+
 /// We don't know much about the actual structs, so better mark them `!Send`,
 /// `!Sync`, `!UnwindSafe`, `!RefUnwindSafe`, `!Unpin` and as mutable behind
 /// shared references.